refactor: enhance TabBar component for dynamic height adjustment
